--- model.cpp
@@ -1,1 +1,0 @@
-#include "model.hpp"
